--- packages/next-intl/src/NextIntlContext.tsx
@@ -1,8 +1,0 @@
-import {createContext} from 'react';
-import Messages from './NextIntlMessages';
-
-const NextIntlContext = createContext<
-  {messages: Messages; locale?: string} | undefined
->(undefined);
-
-export default NextIntlContext;
