--- packages/next-intl/src/NextIntlMessages.tsx
@@ -1,3 +1,0 @@
-type NextIntlMessages = {[id: string]: NextIntlMessages | string};
-
-export default NextIntlMessages;
