--- packages/next-intl/src/TranslationValues.tsx
@@ -1,15 +1,0 @@
-import {ReactNode} from 'react';
-
-type TranslationValues = Record<
-  string,
-  // From IntlMessageFormat#format
-  | string
-  | number
-  | boolean
-  | Date
-  | null
-  | undefined
-  | ((children: ReactNode) => ReactNode)
->;
-
-export default TranslationValues;
