+++ packages/next-intl/src/index.tsx
@@ -1,5 +1,2 @@
 export {default as NextIntlProvider} from './NextIntlProvider';
-export {default as NextIntlMessages} from './NextIntlMessages';
-export {default as useTranslations} from './useTranslations';
-export {default as TranslationValues} from './TranslationValues';
-export {default as useIntl} from './useIntl';
+export * from 'use-intl';
